requirements.txt
# Core dependencies for Scopus ChatBot
gradio>=4.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# For SPECTER model specifically
transformers>=4.21.0
torch>=2.0.0

# Built-in Python modules (no installation needed)
# sqlite3 - Built-in with Python
# json - Built-in with Python
# re - Built-in with Python
# os - Built-in with Python

# Development dependencies (optional)
# jupyter>=1.0.0
# matplotlib>=3.7.0
# plotly>=5.15.0

# Additional dependencies for robustness
scikit-learn>=1.3.0
tqdm>=4.65.0